risk-engine/requirements.txt
# ═══════════════════════════════════════════════════════════════
#  Cosmic Watch — Python Risk Analysis Engine
# ═══════════════════════════════════════════════════════════════

fastapi==0.115.12
uvicorn[standard]==0.34.2
numpy>=1.24,<2.0
scipy>=1.11,<1.14
scikit-learn>=1.3,<1.5
pydantic==2.11.1
pydantic-settings>=2.0,<3.0
httpx==0.28.1
python-socketio[asgi]>=5.11,<6.0
python-dotenv==1.1.0

# ── Scientific Astronomy Libraries ────────────────────────────
astropy>=6.0,<8.0
